move functionality from app.ts to controller.ts
