+++ .eslintrc.js
@@ -11,11 +11,12 @@ module.exports = {
   },
   plugins: ['@typescript-eslint'],
   rules: {
-    indent: ['error', 2],
+    // indent: ['error', 2],
     'linebreak-style': ['error', 'unix'],
     quotes: ['error', 'single'],
     semi: ['error', 'always'],
     '@typescript-eslint/no-require-imports': 'warn',
     '@typescript-eslint/no-var-requires': 'warn',
+    'no-case-declarations': 'off',
   },
 };
